Fixed bug when doing a merge write to a pointer
When a pointer was written through (using a dereference) but it was a
merging write (e.g. when in an array), we were not assigning the new
value back into the domain and the value was lost. Added this case to
the array of pointers tests.

For the two value array we get the wrong result for the final two
assertions as we do not havoc the domain when writing to a pointer which
we don't know what it points to.