Try all rounding modes when rounding mode is unknown
For variable sensitivity domain. If we don't know the rounding mode
(i.e. it's not a constant known value), we can still often
evaluate expressions involving floating points if the results
don't change based on floating point (which is going to be the case
with floating point operations that have exact results or comparisons)