Do not use outdated *-mod.gb file
The regression test wrapper scripts perform a conditional copy of the goto
binary, which is meant to cover the case that an invocation of goto-instrument
does not generate a goto binary. When regression tests are re-run without an
intermittent "clean", the conditional copy was not performed. This may lead to
spurious test failures when goto binary versions change.